Extract app logic and tests
